+++ core/src/components/skeleton-text/skeleton-text.tsx
@@ -1,9 +1,10 @@
-import type { ComponentInterface } from '@stencil/core';
-import { Component, Element, Host, Prop, h } from '@stencil/core';
+import type { ComponentInterface, EventEmitter } from '@stencil/core';
+import { Component, Element, Event, Host, Prop, h } from '@stencil/core';
 import { hostContext } from '@utils/theme';
 
 import { config } from '../../global/config';
 import { getIonMode } from '../../global/ionic-global';
+import type { StyleEventDetail } from '../../interface';
 
 @Component({
   tag: 'ion-skeleton-text',
@@ -17,6 +18,24 @@ export class SkeletonText implements ComponentInterface {
    * If `true`, the skeleton text will animate.
    */
   @Prop() animated = false;
+
+  /**
+   * Emitted when the styles change.
+   * @internal
+   */
+  @Event() ionStyle!: EventEmitter<StyleEventDetail>;
+
+  componentWillLoad() {
+    this.emitStyle();
+  }
+
+  private emitStyle() {
+    const style: StyleEventDetail = {
+      'skeleton-text': true,
+    };
+
+    this.ionStyle.emit(style);
+  }
 
   render() {
     const animated = this.animated && config.getBoolean('animated', true);