+++ packages/angular/test/apps/ng14/e2e/src/angular-version.spec.ts
@@ -1,5 +1,0 @@
-it("should be on Angular 14", () => {
-  cy.visit('/');
-
-  cy.get('ion-title').contains('Angular 14');
-});